feat: create BookmarkView component and refactor Bookmark component structure
