+++ packages/web/components/Bookmark/View.vue
@@ -1,0 +1,58 @@
+<template>
+  <ElmHeading1 text="Bookmarks" />
+  <ElmBlockFallback v-if="bookmarkStore.bookmarkList.length === 0" />
+  <div class="global-fade-in" v-else>
+    <div
+      v-for="b in bookmarkStore.classifiedBookmarkList"
+      :style="{ marginBlock: '1rem' }"
+    >
+      <ElmTag :text="b.tag.name" :color="b.tag.color" />
+      <div class="bookmark-container">
+        <template v-for="bookmark in b.bookmarkList">
+          <ElmBookmarkIcon
+            v-if="bookmark.url != null"
+            :name="bookmark.name ?? bookmark.url ?? ''"
+            :favicon="
+              bookmark.favicon ??
+              'https://www.svgrepo.com/show/197996/internet.svg'
+            "
+            :href="bookmark.url ?? ''"
+          />
+        </template>
+      </div>
+    </div>
+  </div>
+
+  <ElmParagraph v-if="bookmarkStore.error != null">{{
+    bookmarkStore.error
+  }}</ElmParagraph>
+</template>
+
+<script setup lang="ts">
+import {
+  ElmBlockFallback,
+  ElmBookmarkIcon,
+  ElmHeading1,
+  ElmInlineText,
+  ElmParagraph,
+  ElmTag
+} from '@elmethis/core'
+
+const bookmarkStore = useBookmarkStore()
+
+onMounted(async () => {
+  console.log('fetching bookmarks')
+  await bookmarkStore.fetch()
+})
+</script>
+
+<style scoped lang="scss">
+.bookmark-container {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(5rem, 1fr));
+  grid-template-rows: repeat(auto-fill, minmax(2rem, 1fr));
+  grid-template-areas:
+    'header header'
+    'main sidebar';
+}
+</style>
